feat(j): migrate charge.refunded + invoice.* to registry (J-01c-1)
Why:
The 1197-line app/api/stripe/webhook/route.ts is being broken up so
each event handler is one file with its own test surface (J-01).
J-01a scaffolded the registry, J-01b migrated `customer.subscription.*`
+ extracted shared email helpers + `upsertSubscription`. This
iteration migrates `charge.refunded` (3-flow refund processor —
courses / wallet / consultations) plus `invoice.paid` and
`invoice.payment_failed` (advisor-billing delegation + consumer Pro
dunning email). The remaining `checkout.session.completed` (~509 LOC)
is split into J-01c-2 to keep diffs under the cap; only after that
lands does J-01e remove the legacy switch.

Verified callers / context:
- 47 stripe webhook tests pass against the new registry path
  (`__tests__/api/stripe-webhook.test.ts` + `-idempotency.test.ts`)
  with no behavioural changes. The test file references all 4
  migrated event types (`invoice.paid`, `invoice.payment_failed`,
  `charge.refunded`, plus `customer.subscription.*` from J-01b).
- `route.ts` shed `import { handleInvoicePaid, handleInvoicePaymentFailed }
  from "@/lib/advisor-billing"` because those imports moved into the
  new `handlers/invoice.ts`. Greps confirm those symbols are not
  referenced anywhere else in `route.ts`.
- `handlers/index.ts` registers the 3 new handlers; comment placeholder
  for J-01c-2 (checkout.session.completed) remains.
- The legacy `charge.refunded` block returned via `break` (relying on
  the dispatch loop's idempotency stamp); the registry version returns
  `{ status: "done" }` explicitly. The early-return path when
  `payment_intent` is missing is preserved as `return { status: "done" }`
  (legacy was `break`); both produce the same idempotency-row outcome.
- Partial-refund-safe wallet accounting (cumulative `amount_refunded`
  vs prior reversals via `reference_id`) is preserved byte-for-byte.

Idempotency / safety claim:
- All three handlers return `{ status: "done" }` after their work,
  so the dispatch loop marks the idempotency row complete and Stripe
  stops retrying — same terminal state as the legacy `break`.
- `invoice.payment_failed` retains the fire-and-await dunning-email
  semantics (legacy used `await`, not fire-and-forget); a Resend
  outage will block the handler return, leaving the row `processing`
  until Stripe retries — same behaviour as before.
- `charge.refunded` retains the dynamic `import("@/lib/marketplace/wallet")`
  to keep heavy deps off the cold path when refunds aren't wallet-related.

Rollback location:
Revert this commit; the registry will return `{handled:false}` for the
3 newly-registered events and the route's legacy switch arm (still
intact for `checkout.session.completed` until J-01c-2) plus its
`default` arm will pick them up. No partial state.

Refs: #218
Audit: docs/audits/codebase-health-2026-04-24.md §J
Queue: docs/audits/REMEDIATION_QUEUE.md J-01c-1
