+++ app/api/stripe/webhook/route.ts
@@ -1,6 +1,5 @@
 import { getStripe } from "@/lib/stripe";
 import { createAdminClient } from "@/lib/supabase/admin";
-import { handleInvoicePaid, handleInvoicePaymentFailed } from "@/lib/advisor-billing";
 import { NextRequest, NextResponse } from "next/server";
 import type Stripe from "stripe";
 import { logger } from "@/lib/logger";
@@ -147,92 +146,6 @@ export async function POST(request: NextRequest) {
       // Skip the legacy switch — registry owns this event type now.
       // Fall through to the post-switch idempotency stamp below.
     } else switch (event.type) {
-      case "invoice.paid": {
-        const paidInvoice = event.data.object as Stripe.Invoice;
-        const paidPiId =
-          typeof paidInvoice.payment_intent === "string"
-            ? paidInvoice.payment_intent
-            : paidInvoice.payment_intent?.id || null;
-
-        // Check if this is an advisor lead billing invoice
-        if (paidInvoice.metadata?.type === "advisor_lead") {
-          await handleInvoicePaid(paidInvoice.id, paidPiId);
-        }
-
-        log.info("Invoice paid", { invoiceId: paidInvoice.id, customer: paidInvoice.customer });
-        break;
-      }
-
-      case "invoice.payment_failed": {
-        const invoice = event.data.object as Stripe.Invoice;
-
-        // Update advisor billing if applicable
-        if (invoice.metadata?.type === "advisor_lead") {
-          await handleInvoicePaymentFailed(invoice.id);
-        }
-
-        // Notify the subscriber so they can update their card before Stripe's
-        // dunning cycle cancels the subscription. Previously this only
-        // logged a warning, so subscribers with an expired card would
-        // silently slide into past_due → canceled and only find out when
-        // their Pro features stopped working. getSubscription()'s isPro
-        // already excludes past_due, so access is revoked at the check
-        // layer — this handler just adds the user-visible notification.
-        const invCustomerId =
-          typeof invoice.customer === "string" ? invoice.customer : invoice.customer?.id;
-        if (invCustomerId && invoice.metadata?.type !== "advisor_lead") {
-          try {
-            const custAdmin = createAdminClient();
-            const { data: subProfile } = await custAdmin
-              .from("profiles")
-              .select("id")
-              .eq("stripe_customer_id", invCustomerId)
-              .maybeSingle();
-
-            // Tag the subscription with payment_failed_at if the column exists.
-            // Safe under schema drift — errors are logged and swallowed.
-            if (subProfile) {
-              await custAdmin
-                .from("subscriptions")
-                .update({ updated_at: new Date().toISOString() })
-                .eq("stripe_customer_id", invCustomerId);
-            }
-
-            const customer = await getStripe().customers.retrieve(invCustomerId);
-            if (!("deleted" in customer) && customer.email) {
-              const amount = ((invoice.amount_due || 0) / 100).toFixed(2);
-              await sendTransactionalEmail(
-                customer.email,
-                "Action needed: your Invest.com.au Pro payment failed",
-                emailWrapper("Payment Failed ⚠️", "#dc2626", `
-                  <h2 style="margin: 0 0 12px; font-size: 18px; color: #0f172a;">We couldn't process your renewal</h2>
-                  <p style="color: #475569; font-size: 14px; line-height: 1.6; margin: 0 0 16px;">
-                    We tried to charge your card A$${amount} for your Invest.com.au Pro
-                    subscription, but the payment didn't go through.
-                  </p>
-                  <p style="color: #475569; font-size: 14px; line-height: 1.6; margin: 0 0 16px;">
-                    Pro features are temporarily paused until you update your
-                    payment method. We'll retry over the next few days — update
-                    your card now to avoid losing access.
-                  </p>
-                  <div style="text-align: center; margin: 20px 0;">
-                    <a href="${getSiteUrl()}/account" style="display: inline-block; padding: 12px 28px; background: #0f172a; color: #fff; font-weight: 700; font-size: 14px; border-radius: 8px; text-decoration: none;">Update Payment Method →</a>
-                  </div>
-                `),
-              );
-            }
-          } catch (err) {
-            log.error("Payment failed email error", { error: err instanceof Error ? err.message : String(err) });
-          }
-        }
-
-        log.warn("Payment failed for customer", { customer: invoice.customer, invoice: invoice.id });
-        // The subscription.updated webhook will also fire with status 'past_due',
-        // which upsertSubscription handles automatically. Access is revoked
-        // by getSubscription() since past_due is excluded from isPro.
-        break;
-      }
-
       case "checkout.session.completed": {
         const session = event.data.object as Stripe.Checkout.Session;
 
@@ -743,150 +656,11 @@ export async function POST(request: NextRequest) {
         break;
       }
 
-      case "charge.refunded": {
-        const charge = event.data.object as Stripe.Charge;
-        const paymentIntentId =
-          typeof charge.payment_intent === "string"
-            ? charge.payment_intent
-            : charge.payment_intent?.id || null;
-        const refundedAmountCents = charge.amount_refunded;
-        const supabase = createAdminClient();
-
-        if (!paymentIntentId) break;
-
-        // 1. Check if this was a course purchase — revoke access
-        const { data: coursePurchase } = await supabase
-          .from("course_purchases")
-          .select("id, user_id, course_slug, amount_paid")
-          .eq("stripe_payment_id", paymentIntentId)
-          .maybeSingle();
-
-        if (coursePurchase) {
-          await supabase
-            .from("course_purchases")
-            .update({ refunded: true, refunded_at: new Date().toISOString() })
-            .eq("id", coursePurchase.id);
-
-          // Delete associated revenue record
-          await supabase
-            .from("course_revenue")
-            .delete()
-            .eq("purchase_id", coursePurchase.id);
-
-          log.info("Course purchase refunded", { courseSlug: coursePurchase.course_slug, userId: coursePurchase.user_id });
-        }
-
-        // 2. Check if this was a wallet top-up — reverse the credit
-        const { data: walletTxn } = await supabase
-          .from("wallet_transactions")
-          .select("id, broker_slug, amount_cents")
-          .eq("stripe_payment_intent_id", paymentIntentId)
-          .eq("type", "deposit")
-          .maybeSingle();
-
-        if (walletTxn) {
-          try {
-            const { debitWallet } = await import("@/lib/marketplace/wallet");
-
-            // Partial-refund-safe reversal. Stripe sends `charge.refunded`
-            // with `charge.amount_refunded` as the *cumulative* amount
-            // refunded on the charge. If a user has already been partially
-            // refunded (say $50) and we later refund another $50, the
-            // second event has amount_refunded = $100. Naïvely calling
-            // debitWallet($100) would over-reverse by $50.
-            //
-            // Fix: look up prior reversals for this charge (by the
-            // reference_id we set below), sum them, and only reverse the
-            // delta. Also capped against the original top-up amount so a
-            // bug in Stripe's side can never drain more than the deposit.
-            const { data: priorReversals } = await supabase
-              .from("wallet_transactions")
-              .select("amount_cents")
-              .eq("broker_slug", walletTxn.broker_slug)
-              .eq("type", "spend")
-              .eq("reference_type", "stripe_refund")
-              .eq("reference_id", charge.id);
-
-            const alreadyReversedCents = (priorReversals || []).reduce(
-              (sum, r) => sum + (r.amount_cents || 0),
-              0,
-            );
-            const targetReversalCents = Math.min(refundedAmountCents, walletTxn.amount_cents);
-            const deltaCents = targetReversalCents - alreadyReversedCents;
-
-            if (deltaCents <= 0) {
-              log.info("Wallet refund already fully reversed — skipping", {
-                brokerSlug: walletTxn.broker_slug,
-                chargeId: charge.id,
-                alreadyReversedCents,
-                targetReversalCents,
-              });
-            } else {
-              await debitWallet(
-                walletTxn.broker_slug,
-                deltaCents,
-                `Stripe refund reversal — $${(deltaCents / 100).toFixed(2)}`,
-                { type: "stripe_refund", id: charge.id }
-              );
-
-              // Notify broker
-              await supabase.from("broker_notifications").insert({
-                broker_slug: walletTxn.broker_slug,
-                type: "wallet_refund",
-                title: "Wallet Top-Up Reversed",
-                message: `A refund of $${(deltaCents / 100).toFixed(2)} was processed. Your wallet balance has been adjusted.`,
-                link: "/broker-portal/wallet",
-                is_read: false,
-                email_sent: false,
-              });
-
-              log.info("Wallet refund reversed", {
-                brokerSlug: walletTxn.broker_slug,
-                deltaCents,
-                cumulativeRefundedCents: refundedAmountCents,
-              });
-            }
-          } catch (err) {
-            log.error("Wallet refund reversal failed", { error: err instanceof Error ? err.message : String(err) });
-          }
-        }
-
-        // 3. Check if this was a consultation booking — cancel it
-        const { data: booking } = await supabase
-          .from("consultation_bookings")
-          .select("id, user_id, consultation_id")
-          .eq("stripe_payment_id", paymentIntentId)
-          .maybeSingle();
-
-        if (booking) {
-          await supabase
-            .from("consultation_bookings")
-            .update({ status: "refunded", refunded_at: new Date().toISOString() })
-            .eq("id", booking.id);
-
-          log.info("Consultation booking refunded", { bookingId: booking.id });
-        }
-
-        // Audit log
-        await supabase.from("admin_audit_log").insert({
-          action: "stripe_refund",
-          entity_type: "charge",
-          entity_id: charge.id,
-          entity_name: paymentIntentId,
-          details: {
-            amount_refunded_cents: refundedAmountCents,
-            course_purchase_revoked: !!coursePurchase,
-            wallet_reversed: !!walletTxn,
-            consultation_cancelled: !!booking,
-          },
-          admin_email: "stripe_webhook",
-        });
-
-        break;
-      }
-
-      // J-01a: `charge.dispute.created` migrated to the registry.
-      // See `lib/stripe-webhook/handlers/charge-dispute-created.ts`.
+      // J-01a/b/c-1: `charge.dispute.created`, `customer.subscription.*`,
+      // `charge.refunded`, `invoice.paid`, `invoice.payment_failed`
+      // are migrated to the handler registry. See
+      // `lib/stripe-webhook/handlers/`. J-01c-2 will migrate the
+      // remaining `checkout.session.completed` handler.
 
       default:
         // Unhandled event type — log for debugging
